working on gltf meshes

+++ crates/ris_data/src/ecs/mesh.rs
@@ -49,7 +49,7 @@ pub const VERTEX_ATTRIBUTE_DESCRIPTIONS: [vk::VertexInputAttributeDescription; 3
 #[derive(Debug, Default)]
 pub struct Mesh {
     pub vertices: Vec<Vertex>,
-    pub indices: Vec<u32>,
+    pub indices: Vec<u16>,
 }
 
 #[derive(Debug, Default)]
